data-processor/requirements.txt
influxdb-client
paho-mqtt
python-dotenv